requirements.txt
PyYAML==5.1
confluent-kafka==0.11.6
jsonschema==3.0.1
psycopg2-binary==2.7.7
